requirements.txt
google-cloud-vision
streamlit==1.29.0
torch 
torchvision 
torchaudio
transformers
dotenv  